Judge dashboard completed
In input folder verdict will be stored with winner's name as the name of
the file. Client will make an API post request which will contain a file
with arguments between two lawyers. Then client will wait for 10 seconds
and make another API get request to get the verdict generated. The
results will be displayed as winner and verdict.